Fix not logged in flow, Fix moving to center with autocomplete, needed complex fix.
So, autocomplete is annoying. `getPlace()` returns a mostly POD object, except for `geometry`. Geometry contains some getters, `lat()` and `lng()`, to get the location. Except, I can't, or SHOULDN'T store those in redux, since they aren't serializable. So I'm doing an ugly hack. I make a copy of the object before dispatching, pull out the data, remove the getters, merge the extracted data with the rest of the original object, and set that. This is done in a translator function that needs to be manually called, since redux-toolkit also warns on passing the functions into the action itself. So I check in the action whether the original field was removed first and throw an error if it wasn't. I shouldn't hit that path anyways unless I've done something wrong.

ALSO, I don't actually test when not logged in (YET), so I forgot to `skip_before_action :authorized` for a few things. To improve the flow, I needed to not raise `alert`s in a few places (as I was doing by default) when not logged in, and instead render proper errors suggesting the user log in. This may cause some weird issues down the line if I screw up the get_email api later, since it no longer raises top level alert dialog.