OAuth API veification
This is the commit at which point I submitted for OAuth verification. Well, I submitted immediately before this. Google is complaining that my domain isn't verified even though it shows that it is in the search console (it might be because I only did it for the https:// domain), so I'm adding the `meta` tag too, for extraness.

I'm also including the note that I wrote in the 1000 character 'any additional info' field (they say to add anything that might help them) just for my own reference, in case I need to do this again.