Fix for sentry issue, autocomplete returns stub place instead of error
Apparently, the error for autocomplete is a "stub" place with only the name field present, which is just the user input text. So if the user enters some text and then hits enter (which rarely works), then we get a stub/error result. Now, I display a message to the user.