tweak username population logic for async, tweak privacy policy wording
A user reported that sometimes the "loading... (email)" text remains. I strongly suspect this is because that request was a tad slow. This should fix it by checking to see if the username was already populated.

Also, I really don't have any plans to show user emails in the app, so I tweaked the privacy policy for that.