Tweak auth to try and work around sign in issues, update privacy policy
Invalid privacy policy
Based on our review, your privacy policy doesn’t comply with our policy requirements. Please review the Personal and Sensitive Information policy and ensure that you are compliant.

Please add or update your privacy policy. Below is a compliance checklist for privacy policies. Your privacy policy must:
Reference location data and your app’s use and handling of location data
Clearly label that the page is a privacy policy, for example, listed as "privacy policy" in the title or has “privacy” in the URL and body of the page
Be readable in a standard browser without any plug-ins or special handlers (for example, no PDFs, no uncommon MIME types)
Be available on an active, visible URL
Be non-editable
Refer to your app by name
Be linked on your app's Store listing page and/or within your app
Have a URL that links to a single governing privacy policy in your website and/or Play Store listing rather than one that points to multiple privacy policies
Learn more about privacy policy requirements in the Developer Policy Center or in this video.
App Bundle:440010007	Prominent disclosure not found
Your app must display a prominent disclosure through a pop-up alert before your app’s location runtime permission. Based on our review, a prominent disclosure did not appear before the runtime permission.

Please add a prominent disclosure before the runtime permission.

To meet the policy requirements, it’s recommended that you reference one of the following example formats, the second example includes the use of location for ads (choose the most relevant phrasing):
“[This app] collects location data to enable ["feature"], ["feature"], & ["feature"] even when the app is closed or not in use.”
“[This app] collects location data to enable ["feature"], ["feature"], & ["feature"] even when the app is closed or not in use and it is also used to support advertising.”
Example: “Fitness Funds collects location data to enable fitness tracking even when the app is closed or not in use.”

The language in the disclosure MUST include the following elements:
The term “location”
Indication that the nature of usage is in the background by using one of the following phrases “background” / “when the app is closed” / “always in use” / “when the app is not in use”
A list of all the features that use location in the background
If you extend permitted usage to ads, you must include the following: “used to provide ads / support advertising / support ads.” (Choose the most accurate phrasing)
Remember, your prominent disclosure must:
Appear before your app’s location runtime permission
Include an example format listed above
Include any other details necessary to make it clear to the user how and why you are using location in the background. While additional content is permitted, it should not cause the required content to not be immediately visible
Include buttons that are clear to the user if they are approving or denying access. If a user denies access and that would limit the app experience, then make sure to inform them on how to change preferences in the future
Not appear alongside other data privacy disclosures or as part of the privacy policy
Please ensure your prominent disclosure is included in the video required in the Developer console.

Learn more about prominent disclosure requirements in the Developer Policy Center or in this video.