cleanup script to autocheck googleauth version
Ok, so the script seems all good.

The idea is to notify me to check to see if any new updates support the heroku 22 stack, as per issue here: https://github.com/googleapis/google-auth-library-ruby/issues/381

It's an OpenSSL issue. I'm not in a big hurry - this stack will be supported until 	April 2025 - but it would be nice to keep up.