AI onboarding part 2
Excellent work so far. Now, the next thing. I imported the instructions in use from an entirely different repository - the "DeeDee-Prototype" repo. That repo is next to this one in the parent directory of this repo. I manually removed some of the things that might've been confusing for claude code in this repo before the work we're doing now. I'd like you to now examine the patterns in the multiple instructions files in that repo, to see what remains that can be "ported" over to this repo - make a plan on how to proceed thoroughly. I will probably want to re-create some of the memory and context management infrastructure, which is a very complex task, hence wanting you to plan first. Feel free to ultrathink and work as hard as possible for as long as possible on this plan - all the cognitive and planning effort we put in now has a very high ROI on your future abilities and our work together, so I want to maximize our benefits here.